+++ care-reporting/care-reporting-bundle/src/main/java/org/motechproject/care/reporting/mds/measure/PncMotherForm.java
@@ -3,8 +3,8 @@ package org.motechproject.care.reporting.mds.measure;
 import javax.jdo.annotations.Unique;
 
 import org.joda.time.DateTime;
-import org.motechproject.care.reporting.domain.dimension.Flw;
-import org.motechproject.care.reporting.domain.dimension.MotherCase;
+import org.motechproject.care.reporting.mds.dimension.Flw;
+import org.motechproject.care.reporting.mds.dimension.MotherCase;
 import org.motechproject.care.reporting.utils.FormToString;
 import org.motechproject.mds.annotations.Cascade;
 import org.motechproject.mds.annotations.Entity;
@@ -66,7 +66,7 @@ public class PncMotherForm extends Form {
 	private String motherAlive;
 	private String motherChildAlive;
 	private String nextvisittype;
-	private Short numChildren;
+	private int numChildren;
 	private String otherIssues;
 	private String painUrination;
 	private String painfulNipples;
@@ -74,7 +74,7 @@ public class PncMotherForm extends Form {
 	private Integer pnc1DaysLate;
 	private Integer pnc2DaysLate;
 	private Integer pnc3DaysLate;
-	private Short pncVisitNum;
+	private int pncVisitNum;
 	private Boolean ppfpInterest;
 	private String ppiudAbdominalPain;
 	private String ppiudBleeding;
@@ -581,11 +581,11 @@ public class PncMotherForm extends Form {
 	}
 
 	@Field(name = "num_children")
-	public Short getNumChildren() {
+	public int getNumChildren() {
 		return this.numChildren;
 	}
 
-	public void setNumChildren(Short numChildren) {
+	public void setNumChildren(int numChildren) {
 		this.numChildren = numChildren;
 	}
 
@@ -653,11 +653,11 @@ public class PncMotherForm extends Form {
 	}
 
 	@Field(name = "pnc_visit_num")
-	public Short getPncVisitNum() {
+	public int getPncVisitNum() {
 		return this.pncVisitNum;
 	}
 
-	public void setPncVisitNum(Short pncVisitNum) {
+	public void setPncVisitNum(int pncVisitNum) {
 		this.pncVisitNum = pncVisitNum;
 	}
